requirements.txt
av==15.1.0
cython==3.1.2
dask==2025.9.1
easydict==1.13
einops==0.8.1
h5py==3.14.0
ipdb==0.13.13
librosa==0.11.0
loguru==0.7.3
matplotlib==3.10.6
mmcv==2.2.0
MultiScaleDeformableAttention==1.0
numpy==2.3.3
opencv_python==4.11.0.86
pandas==2.3.2
Pillow==11.3.0
pyaudio==0.2.14
pycocotools==2.0.8
pyFFTW==0.15.0
PyYAML==6.0.2
resampy==0.4.3
scikit_learn==1.7.2
scipy==1.16.2
setuptools==75.8.2
soundfile==0.13.1
terminaltables==3.1.10
timm==1.0.19
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchmetrics==1.7.1
torchsummary==1.5.1
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.50.0
Utils==1.0.2
